Allow compilation without X11
