+++ src/output.h
@@ -51,6 +51,16 @@ struct output {
         double (*get_scale)(void);
 };
 
+#ifndef ENABLE_X11
+#define WAYLAND_ONLY 1
+#else
+#define WAYLAND_ONLY 0
+#endif
+
+#if !defined(ENABLE_X11) && !defined(ENABLE_WAYLAND)
+#error "You have to compile at least one output (X11, Wayland)"
+#endif
+
 /**
  * return an initialized output, selecting the correct output type from either
  * wayland or X11 according to the settings and environment.
